refactor(provisioner): split AMI resolution into provisioner_ami (#878)
Clear the S104 (>500 lines) smell on main.py (510 -> 461 lines) by extracting
the cohesive AMI-resolution unit (get_ami_id plus its SSM-path table and
per-invocation cache) into provisioner_ami.py. main re-exports get_ami_id via
__all__, so the main.get_ami_id call sites (terraform_vars) and the
patch("main.get_ami_id") test seam are unchanged. Full provisioner suite:
753 passed, 8 skipped.
