+++ shifter/engine/provisioner/main.py
@@ -79,6 +79,7 @@ __all__ = [
     "_should_run_dc_bootstrap_plan",
     "build_guest_execution_context",
     "generate_presigned_url",
+    "get_ami_id",
     "get_range_availability_zone",
     "has_ngfw_attachment_state",
     "load_range_network_config",
@@ -155,57 +156,6 @@ class DynamicPlan(SetupPlan):
         return {}
 
 
-# SSM parameter paths for AMI IDs (fetched at runtime for latest values)
-_AMI_SSM_PARAMS = {
-    "kali": "/shifter/ami/kali",
-    "victim": "/shifter/ami/ubuntu",
-    "windows": "/shifter/ami/windows",
-    "dc": "/shifter/ami/dc",
-}
-
-# Cache for SSM AMI lookups (cleared per invocation, avoids repeated API calls)
-_ami_cache: dict[str, str] = {}
-
-
-def get_ami_id(ami_type: str) -> str:
-    """Get AMI ID from SSM Parameter Store at runtime.
-
-    This ensures the provisioner always uses the latest AMI IDs without
-    requiring a Terraform apply or ECS task definition update.
-
-    Known types ('kali', 'victim', 'windows', 'dc') use legacy SSM paths.
-    Custom ami_key values resolve to /shifter/ami/<ami_key>.
-
-    Args:
-        ami_type: Known type or custom ami_key (e.g. 'kali', 'windows').
-
-    Returns:
-        AMI ID string
-
-    Raises:
-        ValueError: If SSM parameter not found.
-    """
-    if ami_type in _ami_cache:
-        return _ami_cache[ami_type]
-
-    # Known types use legacy SSM paths; custom keys construct path directly
-    param_path = _AMI_SSM_PARAMS.get(ami_type)
-    if not param_path:
-        param_path = f"/shifter/ami/{ami_type}"
-
-    try:
-        from cloud import get_config_store
-
-        store = get_config_store()
-        ami_id = store.get_parameter(param_path)
-        logger.info("Fetched %s AMI from SSM %s: %s", ami_type, param_path, ami_id)
-        _ami_cache[ami_type] = ami_id
-        return ami_id
-    except Exception as e:
-        # No fallback - fail fast to surface IAM/config issues immediately
-        raise ValueError(f"Failed to get {ami_type} AMI ID from SSM parameter {param_path}: {e}") from e
-
-
 # Default timeout for waiting for NGFW SSH to become available (seconds)
 # PAN-OS boot time is typically 15-25 minutes, but can take longer on first boot
 # 25 minutes
@@ -217,6 +167,7 @@ NGFW_SSH_WAIT_TIMEOUT_DEFAULT = 1500
 # reach in through ``main`` instead of the new sibling module path. They
 # must follow the top-level constants above (hence the per-line E402
 # suppressions) so ``main`` exposes the same surface as the pre-split module.
+from provisioner_ami import get_ami_id  # noqa: E402
 from provisioner_db import (  # noqa: E402
     _append_kwarg_assignment,
     _update_range_config,
